fix: tenantProfile function bug (refactor)

+++ src/controllers/tenant.js
@@ -23,14 +23,6 @@ const tenantProfile = async (req, res) => {
     },
   });
 
-  if (req.user.userId !== tenant.user_id) {
-    return res.status(403).json({ message: "Forbidden" });
-  }
-
-  if (!tenant) {
-    return res.status(404).json({ message: "Tenant not found" });
-  }
-
   return tenant;
 };
 
@@ -38,6 +30,14 @@ const tenantProfile = async (req, res) => {
 const getTenantProfile = async (req, res) => {
   try {
     const tenant = await tenantProfile(req, res);
+    if (req.user.userId !== tenant.user_id) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
+    if (!tenant) {
+      return res.status(404).json({ message: "Tenant not found" });
+    }
+
     return res.status(200).json({ tenant });
   } catch (error) {
     return res.status(500).json({ message: error.message });
@@ -55,7 +55,18 @@ const updateTenantProfile = async (req, res) => {
 
     let updatedTenant = await tenantProfile(req, res);
 
-    if (updatedTenant.avatar_url !== `${process.env.API_DOMAIN}/uploads/avatar/default-avatar.png`) {
+    if (req.user.userId !== updatedTenant.user_id) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+
+    if (!updatedTenant) {
+      return res.status(404).json({ message: "Tenant not found" });
+    }
+
+    if (
+      updatedTenant.avatar_url !==
+      `${process.env.API_DOMAIN}/uploads/avatar/default-avatar.png`
+    ) {
       const oldAvatar = updatedTenant.avatar_url.split("/uploads/avatar/")[1];
       fs.unlinkSync(path.join(__dirname, `../uploads/avatar/${oldAvatar}`));
     }
